+++ include/metashell/engine_clang.hpp
@@ -18,6 +18,8 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #include <metashell/iface/engine.hpp>
+#include <metashell/iface/environment_detector.hpp>
+#include <metashell/data/config.hpp>
 #include <metashell/logger.hpp>
 
 #include <memory>
@@ -25,10 +27,10 @@
 namespace metashell
 {
   std::unique_ptr<iface::engine> create_clang_engine(
-    const std::string& clang_path_,
+    const data::config& config_,
     const std::string& internal_dir_,
     const std::string& env_filename_,
-    const std::vector<std::string>& extra_args_,
+    iface::environment_detector& env_detector_,
     logger* logger_
   );
 }