+++ test/unit/mock_environment_detector.cpp
@@ -26,7 +26,7 @@ mock_environment_detector::mock_environment_detector() :
   _on_windows_called_times(0),
   _on_osx_returns(false),
   _on_osx_called_times(0),
-  _path_of_executable_called_times(0)
+  _directory_of_executable_called_times(0)
 {}
 
 void mock_environment_detector::search_clang_binary_returns(
@@ -101,21 +101,21 @@ int mock_environment_detector::on_osx_called_times() const
   return _on_osx_called_times;
 }
 
-std::string mock_environment_detector::path_of_executable()
+std::string mock_environment_detector::directory_of_executable()
 {
-  ++_path_of_executable_called_times;
-  return _path_of_executable_returns;
+  ++_directory_of_executable_called_times;
+  return _directory_of_executable_returns;
 }
 
-void mock_environment_detector::path_of_executable_returns(
+void mock_environment_detector::directory_of_executable_returns(
   const std::string& result_
 )
 {
-  _path_of_executable_returns = result_;
+  _directory_of_executable_returns = result_;
 }
 
-int mock_environment_detector::path_of_executable_called_times() const
+int mock_environment_detector::directory_of_executable_called_times() const
 {
-  return _path_of_executable_called_times;
+  return _directory_of_executable_called_times;
 }